refactor: consolidate result-table folders into results/ (#691)
Merge the two confusingly-named, near-identical result-table locations
into a single canonical results/ directory:

- result-table.svelte -> results/table-multi-program.svelte
- result-table/ partials -> results/multi-program/ (footer.svelte renamed
  from result-table-footer.svelte)

Update all imports, the calculator page, the moved component tests, and the
component tree in docs/03-architecture.md. Pure move + rename; no logic
changes. The data-testid="result-table" DOM attribute is intentionally kept.
