+++ docs/ai-logs/README.md
@@ -19,6 +19,7 @@ summary table of all sessions.
 
 | File | Date | Topic |
 | ------------------------------------------------------------------ | -------------------------------------------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [2026-06-04-issue-691-consolidate-result-tables.md](2026-06-04-issue-691-consolidate-result-tables.md) | 4 June 2026 | **Consolidate twin result-table folders (#691)** (Claude Opus 4.8 via Claude Code): moved `result-table.svelte` → `results/table-multi-program.svelte` and folded `result-table/` partials into `results/multi-program/` (footer renamed); updated imports, calculator page, moved tests, and `docs/03-architecture.md`. Pure move + rename, no logic changes. |
 | [2026-06-03-issue-631-docs-sync.md](2026-06-03-issue-631-docs-sync.md) | 3 June 2026 | **Sync architecture & specs with Stage 8 state (#631)** (Claude Opus 4.8 via Claude Code): rewrote `docs/03-architecture.md` (project structure, factory-function + orchestrator state topology, layered URL parser `urlv=2`, tabbed `entity-selection/` + `results/table-*` component tree, data-flow diagrams, `<svelte:boundary>` error fallbacks, accessibility); annotated shipped/deferred follow-ups in `entity-selection.md`; verified `calculator.md` already on the unit-anchor strip. |
 | [2026-05-31-inverse-lookups.md](2026-05-31-inverse-lookups.md) | 31 May 2026 | **Second output column for the inverse tabs (#673)** (Claude Opus 4.8 via Claude Code): Range → gains a `→ STP` column (shared `stpOutputUnit` header menu); STP → pairs a `→ Range` column with each energy branch (high-E always, low-E on reveal). Complementary value recovered via a forward `calculate()` at the resolved energies — no new WASM API. State/calc/components + `inverse-lookups.md` v7 / `calculator.md` / wasm contract updated; new unit tests. |
 | [2026-05-30-stp-output-units.md](2026-05-30-stp-output-units.md) | 30 May 2026 | **Stopping-power output unit selector in the Advanced calculator (#670)** (Claude Opus 4.8 via Claude Code): STP column header becomes a dropdown (`keV/µm` · `MeV/cm` · `MeV·cm²/g`) with desktop popover + mobile bottom sheet; shared `sunit=` URL param (single source of truth for calculator + plot; legacy `stp_unit=` still read); cross-page sharing via in-memory state, **no localStorage**; single-entity STP conversion moved to render time. Unit/component/E2E tests added. |
