+++ src/lib/components/results/multi-program/advanced-result-cells.svelte
