+++ src/lib/components/results/multi-program/basic-result-cells.svelte
