+++ src/lib/components/results/multi-program/footer.svelte
