+++ src/lib/components/results/multi-program/multi-entity-cells.svelte
