+++ src/tests/components/table-multi-program-advanced.test.ts
@@ -7,7 +7,7 @@ import { createEntitySelectionState } from "$lib/state/entity-selection.svelte";
 import { createCalculatorState } from "$lib/state/calculator.svelte";
 import { createMultiProgramState, type MultiProgramState } from "$lib/state/multi-program.svelte";
 import { createMultiEntityState } from "$lib/state/multi-entity.svelte";
-import ResultTable from "$lib/components/result-table.svelte";
+import ResultTable from "$lib/components/results/table-multi-program.svelte";
 import type { CalculatorState } from "$lib/state/calculator.svelte";
 import type { EntitySelectionState } from "$lib/state/entity-selection.svelte";
 import type { CalculationResult } from "$lib/wasm/types";
